Fix plugin on Nvim and add g:readline_meta option
The plugin previously did not work in Nvim because the Meta modifier
mappings were not used.  Nvim treats <Esc><Key> as <M-Key> when it is
pressed as a part of a key-chord.

Also add variable `g:readline_meta` which when set to non-zero will load
the Meta modifier mappings into Vim.  This option treated as `0` if not
set, because of the problems with accented characters triggering
mappings.  This option is completely ignored in Nvim.