Exit out of wildmode before cursor move
When wildmenu is active the <left> and <right> commands used for moving
cursor will instead go to previous/next match.  This broke the mappings
M-b, M-f, M-c, C-t and M-t.  To fix this the sequence " \b" is prepended
before each of these mappings, which will exit out of wildmenu if it is
active.

The commands C-b and C-f will also exit out of wildmenu.  The wildmenu
can still be navigated using C-p and C-n.