Change behavior of escape
When g:readline_esc is set to zero (default) the <Esc>x mapping will
only work if <Esc> and x is received simultaneously.  This is the case
when the user presses ALT-x.  When <Esc> is pressed alone it will behave
like default and exit the command-line.  When the user presses an
invalid mapping like <M-a> nothing happens.

When g:readline_esc is set to non-zero it will behave as before were
<Esc> is mapped to <Nop>.  This is to avoid closing the command-line by
accident when pressing an invalid mapping like <M-a>.